Update updateTask Services to Return Updated Task
Previously, the updateTaskStatus and updateTaskTitle services did not
return any data upon a successful update. As a result, code using these
services had to perform an additional database query to access the
updated task.

With this change, both services will now return the updated task,
eliminating the need for an extra database query. This simplifies
the code and enhances efficiency.